分离监听事件
