--- robot/src/main/java/com/mirai/lyf/bot/robot/listener/group/GroupReduceListener.java
@@ -1,7 +1,0 @@
-package com.mirai.lyf.bot.robot.listener.group;
-
-/**
- * 监听群友减少事件
- */
-public class GroupReduceListener {
-}
